+++ RecastDemo/Source/CrowdTool.cpp
@@ -1035,7 +1035,7 @@ void CrowdTool::handleClick(const float* s, const float* p, bool shift)
 			navquery->findNearestPoly(p, halfExtents, &filter, &ref, tgt);
 			if (ref)
 			{
-				unsigned short flags = 0;
+				unsigned int flags = 0;
 				if (dtStatusSucceed(nav->getPolyFlags(ref, &flags)))
 				{
 					flags ^= SAMPLE_POLYFLAGS_DISABLED;
