+++ RecastDemo/Source/NavMeshPruneTool.cpp
@@ -178,7 +178,7 @@ static void disableUnvisitedPolys(dtNavMesh* nav, NavmeshFlags* flags)
 			const dtPolyRef ref = base | (unsigned int)j;
 			if (!flags->getFlags(ref))
 			{
-				unsigned short f = 0;
+				unsigned int f = 0;
 				nav->getPolyFlags(ref, &f);
 				nav->setPolyFlags(ref, f | SAMPLE_POLYFLAGS_DISABLED);
 			}
